Hotfix: trying to create an item of an existing file DELETES THE FILE!
Yeah, I fixed the bug with the "can't delete temp.dat" so good it now deletes even local files. /facepalm

+++ component/backend/tables/item.php
@@ -521,7 +521,7 @@ class ArsTableItem extends FOFTable
 				}
 			}
 
-			if (!empty($filename) && is_file($filename))
+			if (!empty($filename) && is_file($filename) && ($this->type == 'link'))
 			{
 				if (!@unlink($filename))
 				{
